requirements.txt
bcrypt==3.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
google-api-core==2.7.2
google-auth==2.6.5
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
GoogleAppEngineCloudStorageClient==1.9.15.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
paramiko==2.10.3
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
